requirements.txt
openai
python-pptx
streamlit
Pillow
requests